rename rating-description module
