+++ src/utils/get-rating-description.js
